requirements.txt
python-dotenv~=0.19.1
requests~=2.26.0
pygsheets~=2.0.5
python-dateutil~=2.8.2